Adjust QSGTexture comparisonKey type
The original choice was int, simply following textureId(). This was later
deemed insufficient: instead of a GLuint, the value is now often a 64-bit
value (on 64 bit systems), based on a pointer, since the identity of a texture
in the RHI world is the QRhiTexture* itself. In a custom texture implementation
it is likely that the value here is the value of a native object handle, either
a pointer or some 32 or 64 bit integer.

Inspired by the recent QSGTexture::NativeTexture struct change (void* -> quint64),
switch to a qint64 which is big enough to hold all these without truncation.

We choose a signed value here, in order to allow for the following pattern that
is widespread in material compare() implementations:
if (qint64 diff = m_texture->comparisonKey() - other->texture()->comparisonKey())
  return diff;

Fixes: QTBUG-83769

+++ examples/quick/scenegraph/graph/noisynode.cpp
@@ -107,7 +107,7 @@ public:
         if (!state.texture || !other->state.texture)
             return state.texture ? 1 : -1;
 
-        if (int diff = state.texture->comparisonKey() - other->state.texture->comparisonKey())
+        if (qint64 diff = state.texture->comparisonKey() - other->state.texture->comparisonKey())
             return diff;
 
         return 0;
